Integer-based primary keys seem to need to be i32 (INT4 SQL type) for sqlx-postgres compatability

+++ entity/src/lib.rs
@@ -4,4 +4,4 @@ pub mod organization;
 pub mod user;
 
 /// A type alias that represents any Entity's id field data type
-pub type Id = u64;
+pub type Id = i32;
